Move mergeTables
